chore: notifications structure update
