+++ src/services/cleeng.account.service.ts
@@ -220,3 +220,4 @@ const handleErrors = (errors: ApiResponse['errors']) => {
 export const canUpdateEmail = true;
 
 export const canChangePasswordWithOldPassword = false;
+export const subscribeToNotifications = async () => true;
